Stage 4 (2/n): seeds are per-station u64, and the pyo3 boundary exists -- §4.1
THE SEEDING CHANGE, which is what §4.2 has to certify.

  config.seed: i32 -> u64        a station's seed is its identity, and
                                 identities should not be scarce
  DrawSource::for_run(i32)       -> for_station(u64)

The i32 deck seed is why hf_sim.py derives station seeds as
int32(root) ^ stable_hash(name), which lands about half of them negative
and arms the `irand + 8 > 0` jitter gate on the sign of a name hash. In
u64 that footgun cannot be built.

The widening is BIT-EXACT for every deck-sourced seed, deliberately: a
certification is worthless if the stream moved underneath it. main.rs
passes `irand as i64 as u64`, so Modern already computed exactly
seed_from_u64(irand as i64 as u64), and Legacy and Fixture narrow back
with `as i32`, which round-trips a sign-extended value. Confirmed rather
than argued:

  CHEAP PASS -- 22/22 decks within 1e-6, clippy zero, 34+31 tests

So what §4.2 still has to certify is not the widening but the per-station
INDEPENDENCE, which the deck path cannot exercise at all -- nsite=1 was
refused precisely because the Fortran shared one stream across stations.

THE BOUNDARY. src-rust/lib.rs, following site_calculation: Rust is a
numeric kernel, Python is the API. Everything is _-prefixed and takes
arrays; defaults and docstrings live in Python where they are cheapest to
change.

  FaultSegment      keyword-only, so twelve geometry scalars cannot be
                    transposed; grids are (down_dip, along_strike) to
                    match source_modelling.stoch's slip/rise/trup
  SlipModel         segments -> StochModel, derives the aggregates
  VelocityModel1D   arrays -> Moho truncation, one length check for six
                    parallel arrays
  _simulate_stations -> (3, n_station, n_time) f32, cm/s^2

The station loop runs inside py.detach (pyo3 0.29's allow_threads), so it
holds the GIL for approximately none of its runtime and dask's threads
scale across chunks. No internal rayon: with dask on top, two schedulers
would oversubscribe the cores.

No epicentral distance in the return. The Fortran wrote it to stderr
because stderr was the only channel it had to say "I finished", and
hf_sim.py parsed it back and then assigned np.nan over it. A PyErr reports
failure now.

+++ crates/hb_high/src/config.rs
@@ -256,9 +256,18 @@ pub struct HfConfig {
     pub rayset: Vec<RayType>,
     /// `isite_amp != 0`.
     pub site_amp: bool,
-    /// `irand`. Note this is *mutated* by seeding: `init_random_seed` advances it,
-    /// and the advanced value gates the rupture-time jitter.
-    pub seed: i32,
+    /// This station's seed, and the whole of its identity as far as the generator is
+    /// concerned.
+    ///
+    /// `u64` rather than the deck's `i32`. It is per-station, not per-run: each station gets
+    /// an independent PCG stream via [`crate::rng::DrawSource::for_station`], which is what
+    /// makes a batch of stations safe to reorder, subset or resume. The Fortran shared one
+    /// stream across its station loop, which is why `nsite != 1` had to be refused.
+    ///
+    /// The Fortran's `irand` was also *mutated* by seeding — `init_random_seed` advanced it
+    /// and the advanced value gated the rupture-time jitter. That is gone; only
+    /// `HB_LEGACY_SEEDING` still reproduces the sign test.
+    pub seed: u64,
     /// Record length, seconds.
     pub duration: f32,
     /// Sample interval, seconds.
